fix: V4를 ALTER로 변경 + Alert 타입을 V1과 정렬 [S14P31S309-998][S14P31S309-999]
V1 init_schema에 alerts 테이블이 이미 정의되어 있던 것을 검증 중 발견 (BACKEND_STATUS.md의
"alerts 누락" 표기가 outdated였음).

- V4 마이그레이션: CREATE TABLE → ALTER TABLE
  - alert_type VARCHAR(20) → VARCHAR(50) (AGENT_GLUCOSE_HIGH 등 수용)
  - message VARCHAR(100) → VARCHAR(500) (agent 메시지)
  - source 컬럼 추가 (DEFAULT 'rule', CHECK rule|agent)
  - 기존 ck_alerts_type DROP (HIGH/LOW/SOS/WEEKLY_REPORT 한정 — AGENT_* 거부됨)
  - dedup window + unread 인덱스 신설
- Alert 엔티티 타입을 V1과 정렬:
  - id Long → Integer (V1의 INT IDENTITY)
  - weeklyReportId Long → Integer (V1의 INT)
  - latitude/longitude Double → BigDecimal precision(9,6) (V1의 NUMERIC(9,6))
- 전파:
  - AlertRepository<Alert, Long> → <Alert, Integer>
  - AlertCreationService.CreationResult.alertId Long → Integer
  - AgentNotificationItem.alertId Long → Integer
  - AgentNotificationCreateResponse.alertId Long → Integer

검증: docker compose 띄운 뒤 7개 Agent API 모두 정상 동작 확인
- #1~#5 정상/빈 데이터/잘못된 파라미터/인증 누락 케이스
- #6/#7 정상/dedup skip(200)/prefix 위반(400) 케이스
- #3 sleep 7일 평균 산술 정확 (417.1 = (410+425+430+400+415+420+420)/7)

+++ backend/src/main/java/com/ssafy/s309/domain/agent/dto/AgentNotificationCreateResponse.java
@@ -1,8 +1,8 @@
 package com.ssafy.s309.domain.agent.dto;
 
-public record AgentNotificationCreateResponse(Long alertId, Boolean sent, Boolean skipped) {
+public record AgentNotificationCreateResponse(Integer alertId, Boolean sent, Boolean skipped) {
 
-  public static AgentNotificationCreateResponse ofCreated(Long alertId) {
+  public static AgentNotificationCreateResponse ofCreated(Integer alertId) {
     return new AgentNotificationCreateResponse(alertId, true, false);
   }
